+++ AutomatedLabWorker/functions/VirtualMachines/Remove-LWHypervVM.ps1
@@ -39,7 +39,11 @@
     Remove-Item -Path $vmPath -Force -Confirm:$false -Recurse
     
     $vmDescription = Join-Path -Path (Get-Lab).LabPath -ChildPath "$Name.xml"
-    if (Test-Path $vmDescription) {Remove-Item -Path $vmDescription}
+    if (Test-Path $vmDescription) {
+        Remove-Item -Path $vmDescription
+    }
+
+    Remove-LabHypervVmConnectSettingsFile -ComputerName $Name
 
     Write-LogFunctionExit
 }